+++ lib/web_translate_it/util.rb
@@ -23,7 +23,7 @@ module WebTranslateIt
       http = Net::HTTP.new('webtranslateit.com', 443)
       http.use_ssl      = true
       http.verify_mode  = OpenSSL::SSL::VERIFY_NONE
-      http.read_timeout = 10
+      http.read_timeout = 40
       yield http
     end
   end