+++ history.md
@@ -1,3 +1,7 @@
+## Edge
+
+* Increase read timeout to 40s — Required for very large projects.
+
 ## Version 1.4.7 /2010-02-05
 
 * Add deprecation warning for `wti --fetch`, `wti -f`. These commands will be deprecated in favour of `wti pull`.