+++ history.md
@@ -1,8 +1,8 @@
-## Edge
+## Version 1.5.0 /2010-02-19
 
 Warning, some deprecations in this version.
 
-* Remove `wti fetch` and `wti upload`.
+* Remove `wti fetch` and `wti upload`. It was deprecated in 1.4.7 in favour of `wti pull` and `wti push` respectively.
 * `wti push` now only pushes the master language file. Use `wti push --all` to push all or `wti push -l [locale]` to push a specific locale.
 * `wti pull` now only pulls the target language files. Use `wti pull --all` to pull all or `wti pull -l [locale]` to pull a specific locale.
 * Increase read timeout to 40s — Required for very large projects.